Factor out the code to initialize a device for the tests.
